Revert "fix: Resume on all 403 Forbidden responses (#83)"
This reverts commit 8c07830b043376cbd7f5d5c5a45fa08f1afc0687.
